Split base profile file; Added Interfaces
* Renamed class Profile->BaseProfile
* Splitted huge baseprofile file into smaller classes
* Added interfaces for classes
