+++ pyhafas/__init__.py
@@ -2,4 +2,3 @@ from __future__ import annotations
 
 from .client import HafasClient
 from .fptf import Journey, Station
-from .profile import Profile
