+++ pyhafas/profile/__init__.py
@@ -1,3 +1,4 @@
-from .base import Profile, StationBoardRequestType
+from .interfaces import ProfileInterface
+from .base import BaseProfile
 from .db import DBProfile
 from .vsn import VSNProfile
