+++ pyhafas/profile/interfaces/helper/format_products_filter.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class FormatProductsFilterHelperInterface(abc.ABC):
+    @abc.abstractmethod
+    def format_products_filter(self, requested_products: dict) -> dict:
+        pass
