+++ pyhafas/profile/vsn/__init__.py
@@ -1,7 +1,7 @@
-from . import Profile
+from pyhafas.profile import BaseProfile
 
 
-class VSNProfile(Profile):
+class VSNProfile(BaseProfile):
     baseUrl = "https://fahrplaner.vsninfo.de/hafas/mgate.exe"
     defaultUserAgent = "vsn/5.3.1 (iPad; iOS 13.3; Scale/2.00)"
